fix: ignore foreign-injected LMB events (WeChat screenshot Alt+A issue)
Also fix all clippy lints to pass CI (-D warnings):

Mouse proc:
- Add is_foreign_injected_lmb detection: events with flags=0, extra=0
  that are LMB down/up/double pass through CallNextHookEx without remap
  processing, fixing WeChat screenshot Alt+A triggering click instead of
  drag-select.
- Add #[allow(dead_code)] to DragOutput/drag_output/send_fake_drag_move,
  origin_x/origin_y fields in hooks.rs.
- Fix needless_borrow/explicit_auto_deref for Arc<Config> in start() call.
- Fix clone on Copy type for ModifiedDragType.
- Fix unnecessary cast u32->u32 for vkCode.

Formatting: cargo fmt applied to build.rs, gui.rs, scroll/, remap/, remote/,
trackpad_gesture.rs and other files.

Clippy fixes:
- dpi-probe.rs: strip_prefix for ##?# and \??\ prefixes; use array
  instead of closure for split(); fix byte3 = fn_id << 4.
- smoother.rs: use std::f64::consts::PI instead of literal.
- curve.rs: clamp pattern -> .clamp(); remove blank line between doc
  comment and struct; fix assert to use RangeInclusive::contains.
- engine.rs: clamp pattern -> .clamp(); too_many_args allow for
  ScrollAxis::new.
- wheel_tracker.rs: too_many_args allow for both WheelTracker::new.
- window_list.rs: remove duplicate #[allow] attributes.
- remap/mod.rs: use array instead of vec! for INPUT slices; allow
  type_complexity for ClickCycleTracker; use #[allow] for
  too_many_args in send_task_view/send_show_desktop.
- remote.rs: pick_port uses .find() instead of manual loop; SHA-1 loop
  needs needless_range_loop allow.
- trackpad_gesture.rs: fix unreachable 'else if left' (ifs_same_cond);
  int_plus_one -> n > min; add #[allow(dead_code)] to GestureName enum.

Result: 192 tests pass, cargo fmt --check passes, cargo clippy -D warnings passes.

+++ src/device/dpi.rs
@@ -65,7 +65,7 @@ unsafe extern "system" fn monitor_cb(
     let (mut dpi_x, mut dpi_y) = (96u32, 96u32);
     if GetDpiForMonitor(hmon, MDT_RAW_DPI, &mut dpi_x, &mut dpi_y) != 0 && hdc != 0 {
         dpi_x = GetDeviceCaps(hdc, LOGPIXELSX as i32) as u32;
-        dpi_y = GetDeviceCaps(hdc, LOGPIXELSY as i32) as u32;
+        dpi_y = GetDeviceCaps(hdc, LOGPIXELSY) as u32;
     }
     monitors.push(Monitor {
         index: monitors.len(),
